refactor: modified some filenames for uniform file naming
