+++ crates/proof-of-sql/src/arrow/sql_proof_arrow/query_result.rs
